Buying upgrades now changes the text in the shop
There is an issue with the bullet damage as it does not work as intended. Bullet damage is persistent and should not be and my current fix did not fix it but made it so the damage does not increase with upgrades. Also need to adjust the cannon as if you are sprinting and you shoot you can move into the bullet before it shoots out, killing the player.